% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{data}
\name{Iom}
\alias{Iom}
\alias{$.Iom}
\title{Rust class for input-output matrix}
\format{
For some reason, this field introduces itself without asking.
}
\usage{
For some reason, this field introduces itself without asking.
}
\description{
This is the description field!
}
\section{Fields}{

\itemize{
\item \code{name}: A string representing the name of the input-output matrix.
\item \code{intermediate_transactions}: A matrix of intermediate transactions.
\item \code{total_production}: A vector of total production.
\item \code{technical_coefficients_matrix}: A matrix of technical coefficients.
}
}

\section{Methods}{

\itemize{
\item \code{new}:  instantiate new iom class
\item \code{intermediate_transactions}:  Getter for intermediate_transactions matrix.
}


\itemize{
\item \code{compute_technical_coefficients}:  Compute technical coefficients
}
}

\examples{
intermediate_transactions <- matrix(c(1, 2, 3, 4, 5, 6, 7, 8, 9), 3, 3)
total_production <- matrix(c(100, 200, 300), 1, 3)
# instantiate iom object
my_iom <- fiorefactor::iom$new("test", intermediate_transactions, total_production)
# Calculate the technical coefficients
my_iom$compute_technical_coefficients()
# show the technical coefficients
my_iom$technical_coefficients_matrix
}
\keyword{datasets}
