% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.Iom}
\alias{print.Iom}
\title{Print method for objects of class 'Iom'}
\usage{
\method{print}{Iom}(x, ...)
}
\arguments{
\item{x}{An object of class 'Iom'}

\item{...}{Additional arguments passed to print methods}
}
\description{
Print method for objects of class 'Iom'
}
