% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{data}
\name{Test}
\alias{Test}
\title{A test class to demonstrate the use of extendr.}
\usage{
Test$new(name, value)
}
\arguments{
\item{name}{(\code{character})\cr
A string representing the name of the test.}

\item{value}{(\code{numeric})\cr
A numeric value representing the value of the test.}
}
\value{
A new instance of the \code{Test} class.
}
\description{
This is a test struct to check if it's working with multiple structs.
}
\section{Methods}{

\subsection{Method \code{new}}{
Instantiate a new Test object
\subsection{Arguments}{
\describe{
\item{\code{name}}{(\code{character})\cr A string representing the name of the test.}
\item{\code{value}}{(\code{numeric})\cr A numeric value representing the value of the test.}
}}
\subsection{details}
This function creates a new instance of the Test class.
\subsection{return}
A new instance of the Test class.
\subsection{examples}
\preformatted{
Test$new(name = "example", value = 42)
}
}

\subsection{Method \code{get_name}}{
\subsection{details}
This function returns the name of the test.
\subsection{return}
The name of the test.
\subsection{examples}
\preformatted{
test <- Test$new(name = "example", value = 42)
test$get_name()
}
}

\subsection{Method \code{get_value}}{
\subsection{details}
This function returns the value of the test.
\subsection{return}
The value of the test.
\subsection{examples}
\preformatted{
test <- Test$new(name = "example", value = 42)
test$get_value()
}
}
}

\keyword{datasets}
