% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{data}
\name{Iom}
\alias{Iom}
\alias{$.Iom}
\title{Rust class for input-output matrix}
\format{
An object of class \code{environment} of length 7.
}
\usage{
Iom
}
\description{
This is the description
}
\section{Methods}{

\itemize{
\item \code{new}:  instantiate new iom class
\itemize{
\item param: \verb{_name} A string representing the name of the input-output matrix.
\item param: \code{intermediate_transactions} A matrix of intermediate transactions.
\item param: \code{total_production} A vector of total production.
\item return: A new instance of the Iom class.
}
\item \code{intermediate_transactions}:  Getter for intermediate_transactions matrix.
}


\itemize{
\item \code{compute_technical_coefficients}:  Compute technical coefficients
\itemize{
\item param: \code{self} A mutable reference to the Iom instance.
\item return: A new instance of the Iom class with the technical coefficients matrix computed.
}
}
}

\keyword{datasets}
