% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{data}
\name{Iom}
\alias{Iom}
\alias{$.Iom}
\title{Rust class for input-output matrix}
\arguments{
\item{name}{(\code{character})\cr
A string representing the name of the input-output matrix.}

\item{intermediate_transactions}{(\code{matrix})\cr
A matrix of intermediate transactions.}

\item{total_production}{(\code{matrix})\cr A vector of total production.}
}
\value{
A new instance of the \code{Iom} class.
}
\description{
This class represents an input-output matrix, which is a representation of the transactions between different sectors of an economy.
It contains methods to compute the technical coefficients matrix and other related operations.
}
\details{
This paragraph of details is on struct-level.

But hey! This impl-block adds stuff that requires clarification so this paragraph
is on impl-level docs but it's appended to struct docs! Isn't that cool?
}
\section{Methods}{

\subsection{Method \code{new}}{
Instantiate a new Iom object
\subsection{Arguments}{
\describe{
\item{\code{name}}{(\code{character}) A string representing the name of the input-output matrix.}
\item{\code{intermediate_transactions}}{(\code{matrix}) A matrix of intermediate transactions.}
\item{\code{total_production}}{(\code{character}) A vector of total production.}
}}
\subsection{details}{
This function creates a new instance of the Iom class.
}
\subsection{return}{
A new instance of the Iom class.
}
}

\subsection{Method \code{intermediate_transactions}}{
Getter for intermediate_transactions matrix.
}

\subsection{Method \code{total_production}}{
Getter for total_production matrix.
}

\subsection{Method \code{technical_coefficients_matrix}}{
Getter for technical_coefficients_matrix.
}

\subsection{Method \code{leontief_inverse_matrix}}{
Getter for leontief_inverse_matrix.
}


\subsection{Method \code{compute_technical_coefficients}}{
Compute the technical coefficients matrix and populate the \code{technical_coefficients_matrix} field.
\subsection{details}{
It computes the technical coefficientex matrix, a nxn matrix, known as \code{A} matrix, which is the column-wise ratio of intermediate transactions to total production.
}
\subsection{return}{
Self (invisibly)
}
}
}

\examples{
## ---- Method `new` ---- ##
Iom$new(
name = "example",
intermediate_transactions = c(1, 2, 3, 4),
total_production = c(5, 6)
)



## ---- Method `compute_technical_coefficients` ---- ##
iom <- Iom$new(
name = "example",
intermediate_transactions = c(1, 2, 3, 4),
total_production = c(5, 6)
)
iom$compute_technical_coefficients()
iom$technical_coefficients_matrix


}
\keyword{datasets}
